Loading settings from options.ini

+++ Form1.cs
@@ -11,7 +11,7 @@ namespace Repentance_Configuration_Tool
         public string isaacLocation { get; set; }
         public string optionsLocation { get; set; }
 
-        public string options { get; set; }
+        public string[,] optionsValues = new string[39, 2];
         public Form1()
         {
             InitializeComponent();
@@ -29,20 +29,24 @@ namespace Repentance_Configuration_Tool
 
         private void openFileDialog1_FileOk(object sender, CancelEventArgs e)
         {
-           openFileDialog1.OpenFile();
            optionsLocation = openFileDialog1.FileName;
-            try
+            var fileStream = openFileDialog1.OpenFile();
+            string[] fileOptions = new string[fileStream.Length];
+            using (StreamReader reader = new StreamReader(fileStream))
             {
-                options = File.ReadAllText(optionsLocation);
-                int size = options.Length;
-                MessageBox.Show(options);
+                 fileOptions = reader.ReadToEnd().Split('\n');
             }
-            catch (IOException ioExp)
+            fileStream.Close();
+            string[] simpleOption;
+            for (int i = 2; i < fileOptions.Length-2; i++)
             {
-                MessageBox.Show(ioExp.Message);
+                simpleOption = fileOptions[i].Split('=');
+                optionsValues[i-2,0] = simpleOption[0];
+                optionsValues[i-2,1] = simpleOption[1];
+                optionsValues[i-2, 1] = optionsValues[i-2, 1].Replace(",", ".");
             }
-            
-            textBox1.Text.Replace("Options folder not found", "Options folder found");
+            MessageBox.Show(optionsValues[0,0]+":"+optionsValues[0,1]);
+            textBox1.Text = textBox1.Text.Replace("options.ini not found", "options.ini found");
         }
 
         private void openFileDialog2_FileOk(object sender, CancelEventArgs e)
@@ -51,7 +55,8 @@ namespace Repentance_Configuration_Tool
             string isaacFileName = "\\isaac-ng.exe";
             int index = isaacFullLocation.IndexOf(isaacFileName);
             isaacLocation = isaacFullLocation.Replace(isaacFileName,"");
-            textBox1.Text.Replace("Isaac folder not found", "Isaac folder found");
+            button3.Enabled = true;
+            textBox1.Text = textBox1.Text.Replace("Isaac folder not found", "Isaac folder found");
         }
 
         private void domainUpDown1_SelectedItemChanged(object sender, EventArgs e)
@@ -124,13 +129,46 @@ namespace Repentance_Configuration_Tool
                 stringValue = stringValue.Replace(",", ".");
                 stringValue += "000";
             }
-            MessageBox.Show(stringValue);
-            lineChanger("MusicVolume="+stringValue, optionsLocation, 3);
+            //MessageBox.Show(stringValue);
+            try
+            {
+                lineChanger("MusicVolume=" + stringValue, optionsLocation, 3);
+            }
+            catch (NullReferenceException)
+            {
+                MessageBox.Show("options.ini wasn't selected!", "Error!", MessageBoxButtons.OK, MessageBoxIcon.Error);
+            }
+
         }
 
         private void trackBar2_Scroll(object sender, EventArgs e)
         {
             label9.Text = "SFX Volume: " + trackBar2.Value;
+            double SFXvalue = Math.Round(trackBar1.Value % 10.0) / 10;
+            string stringValue = SFXvalue.ToString();
+            if (trackBar1.Value == 0)
+            {
+                stringValue = "0.0000";
+            }
+            else if (trackBar1.Value == 10)
+            {
+                stringValue = "1.0000";
+            }
+            else
+            {
+                stringValue = stringValue.Replace(",", ".");
+                stringValue += "000";
+            }
+            //MessageBox.Show(stringValue);
+            try
+            {
+                lineChanger("SFXVolume=" + stringValue, optionsLocation, 3);
+            }
+            catch (NullReferenceException)
+            {
+                MessageBox.Show("options.ini wasn't selected!", "Error!", MessageBoxButtons.OK, MessageBoxIcon.Error);
+            }
+            //nullreferenceexception
         }
         private void trackBar3_Scroll(object sender, EventArgs e)
         {
@@ -161,9 +199,38 @@ namespace Repentance_Configuration_Tool
 
         private void lineChanger(string newText, string fileName, int line_to_edit)
         {
-            string[] arrLine = File.ReadAllLines(fileName);
-            arrLine[line_to_edit - 1] = newText;
-            File.WriteAllLines(fileName, arrLine);
+            try {
+                string[] arrLine = File.ReadAllLines(fileName);
+                arrLine[line_to_edit - 1] = newText;
+                if (!IsFileLocked(fileName))
+                {
+                    File.WriteAllLines(fileName, arrLine);
+                }  
+            }
+            catch (ArgumentNullException)
+            {
+                MessageBox.Show("options.ini wasn't selected!", "Error!", MessageBoxButtons.OK, MessageBoxIcon.Error);
+            }
+        }
+        public static bool IsFileLocked(string filePath)
+        {
+            bool lockStatus = false;
+            try
+            {
+                using (FileStream fileStream = File.Open(filePath, FileMode.Open, FileAccess.ReadWrite, FileShare.None))
+                {
+                    // File/Stream manipulating code here
+
+                    lockStatus = !fileStream.CanWrite;
+
+                }
+            }
+            catch
+            {
+                //check here why it failed and ask user to retry if the file is in use.
+                lockStatus = true;
+            }
+            return lockStatus;
         }
     }
 }
